requirements.txt
Flask==2.3.3
flask-login==0.6.2
Werkzeug==2.3.7
PyPDF2==3.0.1
google-generativeai==0.4.0
python-dotenv==1.0.0
flask-wtf==1.2.1
Flask-SQLAlchemy==3.0.5
email-validator==2.0.0
Pillow
gunicorn==21.2.0
flask-migrate==4.0.5
requests==2.31.0 
tavily-python
